% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBauerFormat.R
\name{checkBauerFormat}
\alias{checkBauerFormat}
\title{Check whether a dataset fulfills the requirements of the Bauer (2006) format}
\usage{
checkBauerFormat(data)
}
\arguments{
\item{data}{Dataframe object}
}
\description{
Check whether a dataset fulfills the requirements of the Bauer (2006) format
}
