% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.draws_summary}
\alias{.draws_summary}
\title{Summary of posterior distribution}
\usage{
.draws_summary(draws)
}
\arguments{
\item{draws}{Array of posterior draws}
}
\description{
Extracts posterior mean, SD, and 95\% central credible interval from posterior samples
}
