% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_rmmediation.R
\name{report_rmmediation}
\alias{report_rmmediation}
\title{Report results of the repeated-measures mediation model}
\usage{
report_rmmediation(fit, digits = 2)
}
\arguments{
\item{fit}{BRMS model object, result of rmmediation function call}

\item{digits}{Number of reported digits}
}
\description{
Report results of the repeated-measures mediation model
}
