% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_data.R
\name{restructure_data}
\alias{restructure_data}
\title{Restructure data to Bauer (2006) format}
\usage{
restructure_data(X, M, Y, ID)
}
\arguments{
\item{X}{Independent variable (must be numeric)}

\item{M}{Mediator variable (must be numeric)}

\item{Y}{Dependent variable (must be numeric)}

\item{ID}{Subject ID variable}
}
\description{
Restructure data to Bauer (2006) format
}
