% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{rmmediation}
\alias{rmmediation}
\title{Fit Bayesian repeated-measures mediation model}
\usage{
rmmediation(data, ...)
}
\arguments{
\item{data}{Data frame in the Bauer et al. (2006) format. Column names need to contain X, Z, SM, SY, ID.}

\item{...}{Arguments submitted to brm}
}
\description{
Fit Bayesian repeated-measures mediation model
}
